django/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.1
djangorestframework==3.14.0
idna==3.6
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pyotp==2.9.0
pypng==0.20220715.0
pytz==2024.1
qrcode==7.4.2
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.1
